Adds falcoctl 0.9.0 rock
This version of falcoctl is also bundled in the falco image.

Note that Pebble doesn't like it when a process finishes too quickly.
Which is why we're adding a sleep workaround for this rock image, as
it's expected for the image workload to eventually end.

Adds ca-certificates to falcoctl, as it's required by falcoctl to be
able to verify certificates.

Adds falcoctl usage to integration test.