Adds falco-driver-loader 0.38.2 rock
The falco-driver-loader is just the falco image with a different
entrypoint.

Note that Pebble doesn't like it when a process finishes too quickly.
Which is why we're adding a sleep workaround for this rock image, as
it is expected for the image workload to eventually end.

Adds falco-driver-loader usage to the integration test. We also need to
set the "driver.kind=modern_ebpf" Helm chart option, in order for the
falco-driver-loader init container to properly run. By default, it will
try to autodetect the right driver, but it will fail when trying to use
the kmod driver (insmod will fail with "Operation not permitted"), but
the script exit code is still 0, meaning that the init container "successfully"
finishes, which results in the workload container to not start properly.