Adds falco 0.38.2 rock (#2)
On ubuntu-24.04, we have gcc 13, and abseil (grpc's dependency) fails to build with
this version of gcc. Thus, we're building with gcc 12. [1]

In the original Dockerfile, falco is installed using apt. We're creating a .deb file
for Falco and then unpack it into the part install folder. This makes it easier add
all the necessary files into the rock.

[1]: https://github.com/apache/arrow/issues/36969