Adds falcosidekick to the integration test
The falcosidekick helm chart doesn't pass any arguments to
falcosidekick, but we've set the --help as the default argument, meaning
that it won't start up as intended.

falcosidekick helm chart also creates a deployment with
readOnlyRootFilesystem=True set, which means that Pebble won't be able
to copy its necessary files. We can use the Pebble mutating webhook to
solve this issue.