falco: Adds jammy packages in the building stages (#16)
We're currently building rocks using LXD, meaning that they LXDs will
share the host's kernel. In the GitHub action, we're building on
ubuntu 22.04 nodes.

In the falco and falco-driver-loader rocks, we have set the base rocks
to ubuntu 24.04. The building LXD container will have the 24.04 apt
package repositories, though we're going to need 22.04's repositories in
order to install the right linux-headers, linux-tools, and
linux-cloud-tools (based on uname -r). Otherwise, falco will fail to
build.